화면은 이번 개명에 안 걸린다 — 대신 조용한 구멍 둘을 찾았다
web 이 화면의 입력을 전수 조사했다. 경로는 WS(9900/ws)와 data_log.json 폴링 둘뿐이고
서버 로그 파일은 어느 경로로도 읽지 않는다. 바뀌는 이름은 소크 요약 로그 줄(server.cpp:808)
이라 화면이 읽지 않는다 — 1단계(구/신 동시 수용)를 넣을 대상이 없다.

모르는 프레임 타입과 모르는 error.code 를 조용히 흘려보내는 설계라 새 필드가 생겨도
안 깨진다는 것도 확인됐다.

찾은 구멍 둘:
- last_frame_ts 를 화면이 읽는데 WS 스냅샷에는 그 키가 없다(server.cpp:1174-1178).
  항상 "모름"으로 떨어진다 — 신선도 표시가 주 경로에서 작동하지 않는다는 뜻이다.
- 최상위 비트열이 아니라 slots[] 만 읽으므로, slots[] 를 빼는 변경이 오면 오류 없이
  10칸 전부 "상태 미상"이 된다. 조용히 접히는 실패다.

그리고 개명이 전선 JSON 키(device.online / device_offline)까지 오는지를 socket 에게
직접 물었다(REQ-0123). 추측해서 넣으면 정의가 어긋난다는 판단이다.

+++ 조별과제샘플/client.ino
@@ -2174,7 +2174,10 @@ static void statusTick(unsigned long now) {
 #if DEBUG
     // 원인은 **진단으로만** 남긴다 — 원인별 제어 경로를 만들면 다시 "아는 실패만" 막게 된다.
     Serial.print(F("[NET] ★ 정지 감지: "));
-    Serial.print((unsigned long)(now - lastTxOkAt));
+    // ⚠ `(uint32_t)` 캐스트는 실기에서는 무의미하지만(unsigned long 이 곧 32비트)
+    //   호스트 테스트에서는 필요하다 — 64비트 `unsigned long` 과 32비트 `lastTxOkAt` 을
+    //   섞으면 언더플로가 감싸이지 않아 1.8e19 같은 값이 찍힌다. 판정식과 같은 폭으로 맞춘다.
+    Serial.print((unsigned long)(uint32_t)(now - lastTxOkAt));
     Serial.print(F("ms 동안 한 줄도 못 나갔다 (busy "));
     Serial.print(stallBusy);
     Serial.print(F(" / 무응답 ")); Serial.print(stallTimeout);
